+++ app/validators/subscribers/create_validator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Subscribers
+  class CreateValidator < ApplicationValidator
+    include Deps[contract: 'contracts.subscribers.create']
+  end
+end